--- a/src/OpenTelemetry/Metrics/AggregatorStore.cs
+++ b/src/OpenTelemetry/Metrics/AggregatorStore.cs
@@ -25,14 +25,15 @@ namespace OpenTelemetry.Metrics
 {
     internal sealed class AggregatorStore
     {
-        private static readonly ObjectArrayEqualityComparer ObjectArrayComparer = new ObjectArrayEqualityComparer();
         private readonly object lockZeroTags = new object();
         private readonly HashSet<string> tagKeysInteresting;
         private readonly int tagsKeysInterestingCount;
 
-        // Two-Level lookup. TagKeys x [ TagValues x Metrics ]
-        private readonly ConcurrentDictionary<string[], ConcurrentDictionary<object[], int>> keyValue2MetricAggs =
-            new ConcurrentDictionary<string[], ConcurrentDictionary<object[], int>>(new StringArrayEqualityComparer());
+        private readonly ConcurrentDictionary<Tags, Tags> sortedTagsDictionary =
+            new ConcurrentDictionary<Tags, Tags>();
+
+        private readonly ConcurrentDictionary<Tags, int> tagsToMetricPointIndexDictionary =
+            new ConcurrentDictionary<Tags, int>();
 
         private readonly AggregationTemporality temporality;
         private readonly string name;
@@ -178,27 +179,46 @@ namespace OpenTelemetry.Metrics
         [MethodImpl(MethodImplOptions.AggressiveInlining)]
         private int LookupAggregatorStore(string[] tagKeys, object[] tagValues, int length)
         {
-            int aggregatorIndex;
-            string[] seqKey = null;
+            var givenTags = new Tags(tagKeys, tagValues);
 
-            // GetOrAdd by TagKeys at 1st Level of 2-level dictionary structure.
-            // Get back a Dictionary of [ Values x Metrics[] ].
-            if (!this.keyValue2MetricAggs.TryGetValue(tagKeys, out var value2metrics))
+            // We only need to sort if there is more than one Tag Key.
+            if (!this.sortedTagsDictionary.TryGetValue(givenTags, out var sortedTags))
             {
                 // Note: We are using storage from ThreadStatic, so need to make a deep copy for Dictionary storage.
-                seqKey = new string[length];
-                tagKeys.CopyTo(seqKey, 0);
+                var givenKeys = new string[length];
+                tagKeys.CopyTo(givenKeys, 0);
 
-                value2metrics = new ConcurrentDictionary<object[], int>(ObjectArrayComparer);
-                if (!this.keyValue2MetricAggs.TryAdd(seqKey, value2metrics))
+                var givenValues = new object[length];
+                tagValues.CopyTo(givenValues, 0);
+
+                givenTags = new Tags(givenKeys, givenValues);
+
+                string[] sortedTagKeys;
+                object[] sortedTagValues;
+
+                if (length > 1)
                 {
-                    this.keyValue2MetricAggs.TryGetValue(seqKey, out value2metrics);
+                    // Create a new array for the sorted Tag keys.
+                    sortedTagKeys = new string[length];
+                    tagKeys.CopyTo(sortedTagKeys, 0);
+
+                    // Create a new array for the sorted Tag values.
+                    sortedTagValues = new object[length];
+                    tagValues.CopyTo(sortedTagValues, 0);
+
+                    Array.Sort(sortedTagKeys, sortedTagValues);
                 }
+                else
+                {
+                    sortedTagKeys = givenKeys;
+                    sortedTagValues = givenValues;
+                }
+
+                sortedTags = new Tags(sortedTagKeys, sortedTagValues);
+                this.sortedTagsDictionary.TryAdd(givenTags, sortedTags);
             }
 
-            // GetOrAdd by TagValues at 2st Level of 2-level dictionary structure.
-            // Get back Metrics[].
-            if (!value2metrics.TryGetValue(tagValues, out aggregatorIndex))
+            if (!this.tagsToMetricPointIndexDictionary.TryGetValue(sortedTags, out var aggregatorIndex))
             {
                 aggregatorIndex = this.metricPointIndex;
                 if (aggregatorIndex >= this.maxMetricPoints)
@@ -210,12 +230,12 @@ namespace OpenTelemetry.Metrics
                     return -1;
                 }
 
-                lock (value2metrics)
+                lock (this.tagsToMetricPointIndexDictionary)
                 {
                     // check again after acquiring lock.
-                    if (!value2metrics.TryGetValue(tagValues, out aggregatorIndex))
+                    if (!this.tagsToMetricPointIndexDictionary.TryGetValue(sortedTags, out aggregatorIndex))
                     {
-                        aggregatorIndex = Interlocked.Increment(ref this.metricPointIndex);
+                        aggregatorIndex = ++this.metricPointIndex;
                         if (aggregatorIndex >= this.maxMetricPoints)
                         {
                             // sorry! out of data points.
@@ -225,24 +245,14 @@ namespace OpenTelemetry.Metrics
                             return -1;
                         }
 
-                        // Note: We are using storage from ThreadStatic, so need to make a deep copy for Dictionary storage.
-                        if (seqKey == null)
-                        {
-                            seqKey = new string[length];
-                            tagKeys.CopyTo(seqKey, 0);
-                        }
-
-                        var seqVal = new object[length];
-                        tagValues.CopyTo(seqVal, 0);
-
                         ref var metricPoint = ref this.metricPoints[aggregatorIndex];
                         var dt = DateTimeOffset.UtcNow;
-                        metricPoint = new MetricPoint(this.aggType, dt, seqKey, seqVal, this.histogramBounds);
+                        metricPoint = new MetricPoint(this.aggType, dt, sortedTags.Keys, sortedTags.Values, this.histogramBounds);
 
                         // Add to dictionary *after* initializing MetricPoint
                         // as other threads can start writing to the
                         // MetricPoint, if dictionary entry found.
-                        value2metrics.TryAdd(seqVal, aggregatorIndex);
+                        this.tagsToMetricPointIndexDictionary.TryAdd(sortedTags, aggregatorIndex);
                     }
                 }
             }
@@ -355,11 +365,6 @@ namespace OpenTelemetry.Metrics
 
             storage.SplitToKeysAndValues(tags, tagLength, out var tagKeys, out var tagValues);
 
-            if (tagLength > 1)
-            {
-                Array.Sort(tagKeys, tagValues);
-            }
-
             return this.LookupAggregatorStore(tagKeys, tagValues, tagLength);
         }
 
@@ -386,11 +391,6 @@ namespace OpenTelemetry.Metrics
             {
                 this.InitializeZeroTagPointIfNotInitialized();
                 return 0;
-            }
-
-            if (actualLength > 1)
-            {
-                Array.Sort(tagKeys, tagValues);
             }
 
             return this.LookupAggregatorStore(tagKeys, tagValues, actualLength);