… src/OpenTelemetry/Metrics/ObjectArrayEqualityComparer.cs
… src/OpenTelemetry/Metrics/StringArrayEqualityComparer.cs