Fixes item 1 of #2374

## Changes

- In order to avoid sorting the tags on every update call, we keep track of the tag combinations that we sorted so far.
- This mapping of user provided tag combinations to their respective sorted tag combination is maintained in `ConcurrentDictionary<Tags, Tags> sortedTagsDictionary`.
- On every update, we first check if we have sorted the user provided tags combination before.
  + If yes, we then lookup the MetricPoint index for the sorted combination in `ConcurrentDictionary<Tags, int> tagsToMetricPointIndexDictionary`.
  + If no, then we sort the user provided tag combination and add the mapping to the `sortedTagsDictionary`
- The logic for `tagsToMetricPointIndexDictionary` is comparable to the previously used `value2metrics` dictionary. The only difference is that we now use the composite key of tag keys and tag values as the `Key` for the dictionary.
- `Tags` is a struct that contains `string[]` keys and `object[]` values. It represents a tag key-value combination. This is used as the key for both the new dictionaries.

Lessons learnt from #2777: We have to sort the key and values together and their mapping information has to be stored together as well.