@utpilla 2022-01-22
## Benchmarks
#### There is up to ~46% improvement in Perf for higher number of Tags

// * Summary *

BenchmarkDotNet=v0.13.1, OS=Windows 10.0.22000
Intel Core i7-9700 CPU 3.00GHz, 1 CPU, 8 logical and 8 physical cores
.NET SDK=6.0.101
  [Host]     : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT
  DefaultJob : .NET 6.0.1 (6.0.121.56705), X64 RyuJIT

### main

|                    Method | AggregationTemporality |      Mean |     Error |    StdDev | Allocated |
|-------------------------- |----------------------- |----------:|----------:|----------:|----------:|
|            CounterHotPath |             Cumulative |  15.82 ns |  0.078 ns |  0.069 ns |         - |
| CounterWith1LabelsHotPath |             Cumulative |  71.28 ns |  0.270 ns |  0.252 ns |         - |
| CounterWith3LabelsHotPath |             Cumulative | 390.50 ns |  1.550 ns |  1.450 ns |         - |
| CounterWith5LabelsHotPath |             Cumulative | 594.25 ns | 11.624 ns | 13.386 ns |         - |
| CounterWith6LabelsHotPath |             Cumulative | 720.17 ns |  7.979 ns |  6.663 ns |         - |
| CounterWith7LabelsHotPath |             Cumulative | 768.72 ns |  5.820 ns |  4.544 ns |         - |
|            CounterHotPath |                  Delta |  15.68 ns |  0.097 ns |  0.086 ns |         - |
| CounterWith1LabelsHotPath |                  Delta |  68.67 ns |  0.233 ns |  0.195 ns |         - |
| CounterWith3LabelsHotPath |                  Delta | 375.43 ns |  3.357 ns |  3.140 ns |         - |
| CounterWith5LabelsHotPath |                  Delta | 588.71 ns |  5.730 ns |  5.360 ns |         - |
| CounterWith6LabelsHotPath |                  Delta | 683.18 ns |  3.388 ns |  3.169 ns |         - |
| CounterWith7LabelsHotPath |                  Delta | 782.50 ns |  5.883 ns |  5.503 ns |         - |


### With the new changes

|                    Method | AggregationTemporality |      Mean |    Error |   StdDev | Allocated |
|-------------------------- |----------------------- |----------:|---------:|---------:|----------:|
|            CounterHotPath |             Cumulative |  15.61 ns | 0.046 ns | 0.039 ns |         - |
| CounterWith1LabelsHotPath |             Cumulative |  87.49 ns | 0.626 ns | 0.585 ns |         - |
| CounterWith3LabelsHotPath |             Cumulative | 219.34 ns | 1.950 ns | 1.824 ns |         - |
| CounterWith5LabelsHotPath |             Cumulative | 344.24 ns | 1.789 ns | 1.673 ns |         - |
| CounterWith6LabelsHotPath |             Cumulative | 388.55 ns | 1.570 ns | 1.392 ns |         - |
| CounterWith7LabelsHotPath |             Cumulative | 434.44 ns | 3.176 ns | 2.971 ns |         - |
|            CounterHotPath |                  Delta |  15.64 ns | 0.078 ns | 0.069 ns |         - |
| CounterWith1LabelsHotPath |                  Delta |  88.44 ns | 0.554 ns | 0.463 ns |         - |
| CounterWith3LabelsHotPath |                  Delta | 220.13 ns | 0.982 ns | 0.871 ns |         - |
| CounterWith5LabelsHotPath |                  Delta | 335.49 ns | 1.135 ns | 1.006 ns |         - |
| CounterWith6LabelsHotPath |                  Delta | 392.99 ns | 2.364 ns | 2.096 ns |         - |
| CounterWith7LabelsHotPath |                  Delta | 430.89 ns | 1.063 ns | 0.995 ns |         - |