@utpilla 2022-01-22
## Stress Tests
- There is an **increase of ~7M loops/sec** from 14.7M in the main branch to 21.5M with the new changes
- The various memory size numbers are comparable for both the branches


### main
![image](https://user-images.githubusercontent.com/66651184/150623480-e2185d4c-0c19-4b11-a8af-87df792671f6.png)


### With the new changes
![image](https://user-images.githubusercontent.com/66651184/150623491-83849751-0270-4c9b-b33f-fb62ba050eae.png)