@utpilla 2022-02-02
> Looking at the `LookupAggregatorStore` method, the `if (length > 1)` and `else` bodies are almost identical...
> 
> * Lines 185-244 Perform the sort, and then add both `givenTags` and `sortedTags`.
> * Lines 251-294 Are only adding the `givenTags`.
> 
> If it's important to keep these bodies consistent, consider refactoring this out to a helper method.

I have created the issue #2843 to track this.